requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
